+++ tamago-storage/src/main/kotlin/tamago/server/storage/monster/entity/MonsterUnlockPolicyEntity.kt
@@ -14,7 +14,7 @@ import jakarta.persistence.JoinColumn
 import jakarta.persistence.ManyToOne
 import jakarta.persistence.Table
 import tamago.server.core.monster.domain.enum.MonsterRuleType
-import tamago.server.storage.common.entity.BaseTimeEntity
+import tamago.server.storage.support.BaseTimeEntity
 
 @Entity
 @Table(name = "t_monster_unlock_policy")
